rename config stuff
