+++ gl2f/__init__.py
@@ -1,3 +1,3 @@
 from .core.lister import list_contents
 from .core.board import content_url
-from .core import board, member
+from .core import board, member, config
