+++ gl2f/configurator.py
@@ -1,0 +1,73 @@
+from .core import config
+import os
+
+# editors
+def make_string(message):
+	return lambda:input(f'{message}: ')
+
+def make_strings(message, sep=' '):
+	return lambda:input(f'{message}: ').split(sep)
+
+def make_numbers(message, length=1):
+	return lambda:tuple(map(int, input(f'{message}: ').split(maxsplit=length)[:length]))
+
+def make_number(message):
+	return lambda:int(input(f'{message}: ').split(maxsplit=1)[0])
+
+def get_editors():
+	return {
+		'max-image-size': make_numbers('width height in pixels', 2),
+		'serve-port': make_number('port'),
+		'host-name': make_string('host name'),
+	}
+
+def edit():
+	from .ayame import terminal as term
+
+	editors = get_editors()
+	data = config.load()
+
+	for key in term.selected(list(editors.keys()), lambda k:f'{k}: {data.get(k, f"{str(config.data[k])} (default)")}'):
+		try:
+			data[key] = editors[key]()
+		except Exception as e:
+			print(term.mod('error', term.color('red')), e)
+
+	config.save(data)
+
+
+def view():
+	import json
+	path = config.filepath()
+	if not os.path.isfile(path):
+		print('file does not exist.')
+		return
+
+	with open(path, encoding='utf-8') as f:
+		print(f.read())
+
+def create():
+	path = config.filepath()
+	if os.path.isfile(path):
+		print('file already exits. abort.')
+		return
+
+	config.save(config.default())
+	print(f'created "{path}".')
+
+def path():
+	path = config.filepath()
+	print(path if os.path.isfile(path) else 'file does not exist')
+
+def add_to():
+	return 'gl2f', 'config'
+
+def add_args(parser):
+	sub = parser.add_subparsers()
+
+	sub.add_parser('create').set_defaults(handler=lambda _:create())
+	sub.add_parser('path').set_defaults(handler=lambda _:path())
+	sub.add_parser('view').set_defaults(handler=lambda _:view())
+	sub.add_parser('edit').set_defaults(handler=lambda _:edit())
+
+	return sub
